- one lod to 16 combination of index buffer

+++ src/render/src/sge_render/command/RenderCommand.h
@@ -68,7 +68,7 @@ public:
 
 	RenderPrimitiveType		primitive = RenderPrimitiveType::None;
 	const VertexLayout* pVertexLayout = nullptr;
-	RenderDataType			indexType = RenderDataType::UInt16;
+	RenderDataType			indexType = RenderDataType::UInt32;
 
 	SPtr<RenderGpuBuffer>	spVertexBuffer;
 	SPtr<RenderGpuBuffer>	spIndexBuffer;
